Added: Feedback notice in dashboard page.
Added: Feedback notice in dashboard page.
#278
checks for transient and an option status for 'addonify_wishlist_plugin_review_status' , if transient is set or status is 'reviewed', do not show admin message.
On I already did, status is set and message is not shown again. On maybe later, transient is set for 3 days